# Register more infer requests to improve concurrency