This PR is in conjunction with https://github.com/triton-inference-server/client/pull/55 to improve single process GRPC throughput. After the change, the single perf analyzer performance is comparable with two perf analyzer sending requests to the same server (total of ~6300 infer / sec). Note that there is still abnormal grpc time compare to server-side latency which suggests long network overhead, networking analysis is needed to determine if it can be further improved

before:
```
install/bin/perf_analyzer -m batching --concurrency-range=1024 -i grpc -a -v -p 10000
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 10000 msec
  Using asynchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1024
  Pass [1] throughput: 4315.2 infer/sec. Avg latency: 236408 usec (std 65277 usec)
  Pass [2] throughput: 4466.6 infer/sec. Avg latency: 229143 usec (std 58849 usec)
  Pass [3] throughput: 4572.1 infer/sec. Avg latency: 223281 usec (std 78012 usec)
  Client: 
    Request count: 45721
    Throughput: 4572.1 infer/sec
    Avg latency: 223281 usec (standard deviation 78012 usec)
    p50 latency: 243031 usec
    p90 latency: 267150 usec
    p95 latency: 275639 usec
    p99 latency: 435332 usec
    Avg gRPC time: 223703 usec (marshal 119 usec + response wait 223583 usec + unmarshal 1 usec)
  Server: 
    Inference count: 55845
    Execution count: 54387
    Successful request count: 54387
    Avg request latency: 111 usec (overhead 41 usec + queue 63 usec + compute input 3 usec + compute infer 0 usec + compute output 4 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1024, throughput: 4572.1 infer/sec, latency 223281 usec
```

after:
```
../install/bin/perf_analyzer -m batching --concurrency-range=1024 -i grpc -a -v -p 10000
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 10000 msec
  Using asynchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1024
  Pass [1] throughput: 6331.6 infer/sec. Avg latency: 160157 usec (std 38960 usec)
  Pass [2] throughput: 6351.6 infer/sec. Avg latency: 159802 usec (std 29872 usec)
  Pass [3] throughput: 6304.6 infer/sec. Avg latency: 160978 usec (std 32013 usec)
  Client: 
    Request count: 63046
    Throughput: 6304.6 infer/sec
    Avg latency: 160978 usec (standard deviation 32013 usec)
    p50 latency: 159044 usec
    p90 latency: 197871 usec
    p95 latency: 215177 usec
    p99 latency: 250667 usec
    Avg gRPC time: 161364 usec (marshal 86 usec + response wait 161277 usec + unmarshal 1 usec)
  Server: 
    Inference count: 75572
    Execution count: 68081
    Successful request count: 68081
    Avg request latency: 251 usec (overhead 75 usec + queue 165 usec + compute input 5 usec + compute infer 0 usec + compute output 6 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1024, throughput: 6304.6 infer/sec, latency 160978 usec
```