@tanmayv25 2022-01-13
@GuanLuo Can you also experiment with multiple handler threads?
Just create one more handler like this on the same completion_queue(model_infer_cq_). All the request state is held in the State object and might give better concurrency?
https://github.com/triton-inference-server/server/blob/main/src/servers/grpc_server.cc#L4172-L4178