@GuanLuo 2022-01-13
I have experimented with multi handler thread (creating multiple handler object) but the effect is just the same as what this change brings. So I prefer this one as it requires less threads